% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{as.data.frame.openaq_countries_list}
\alias{as.data.frame.openaq_countries_list}
\title{Method for converting openaq_countries_list to data frame.}
\usage{
\method{as.data.frame}{openaq_countries_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of countries as returned from list_countries.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the countries results, with the following
columns:
\describe{
\item{id}{Numeric. The countries identifier}
\item{name}{Character. Then English name of the country.}
\item{code}{Character. The ISO-3166 Alpha 2 identifier for the country.}
\item{datetime_first}{POSIXct. The datetime of the first measurement
value available in the country.}
\item{datetime_last}{POSIXct. The datetime of the last measurement
value available in the country.}
\item{parameter_ids}{Character. A comma delimited list of parameter ids
that are measured within the country.}
\item{parameter_names}{Character. a comma delimited list of parameter
names and their units that are measured within the country.}
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_countries_list}.
}
\description{
Method for converting openaq_countries_list to data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
countries <- list_countries(as_data_frame = FALSE)
as.data.frame(countries)
\dontshow{\}) # examplesIf}
}
