% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{as.data.frame.openaq_instruments_list}
\alias{as.data.frame.openaq_instruments_list}
\title{Method for converting openaq_instruments_list to data frame.}
\usage{
\method{as.data.frame}{openaq_instruments_list}(data, ...)
}
\arguments{
\item{data}{A list of countries as returned from list_instruments.}

\item{...}{Other options.}
}
\description{
Method for converting openaq_instruments_list to data frame.
}
\examples{
\dontrun{
instruments <- list_instruments(as_data_frame = FALSE)
openaq_instruments_list.as.data.frame(instruments)
}
}
