% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{as.data.frame.openaq_latest_list}
\alias{as.data.frame.openaq_latest_list}
\title{Method for converting openaq_latest_list to data frame.}
\usage{
\method{as.data.frame}{openaq_latest_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of latest measurements as returned from list_location_latest
or list_parameter_latest.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the latest results, with the following
columns:
\describe{
\item{sensors_id}{Numeric. The sensors identifier. }
\item{locations_id}{Numeric. The locations identifier.}
\item{value}{Numeric. The measurement value.}
\item{datetime_local}{POSIXct. The datetime of the measurement value,
in local time}
\item{datetime_utc}{POSIXct. The datetime of the measurement value,
in UTC time}
\item{latitude}{Numeric. The latitude, geographic Y, value for the
measurement.}
\item{longitude}{Numeric.  The longitude, geographic X, value for the
measurement.}
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_latest_list}.
}
\description{
Method for converting openaq_latest_list to data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
latest <- list_location_latest(2178, as_data_frame = FALSE)
as.data.frame(latest)
\dontshow{\}) # examplesIf}
}
