% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{as.data.frame.openaq_licenses_list}
\alias{as.data.frame.openaq_licenses_list}
\title{Method for converting openaq_licenses_list to data frame.}
\usage{
\method{as.data.frame}{openaq_licenses_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of licenses as returned from list_licenses.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the licenses results, with the following
columns:
\describe{
\item{id}{Numeric. The licenses identifier. }
\item{name}{Character. The license name. }
\item{commercial_use_allowed}{Logical. Indicates whether commercial use
is allowed under the license terms. }
\item{attribution_required}{Logical. Indicates whether attribution
is required under the license terms. }
\item{share_alike_required}{Logical. Indicates whether share-alike
is required under the license terms.}
\item{modification_allowed}{Logical. Indicates whether modification
is allowed under the license terms. }
\item{redistribution_allowed}{Logical. Indicates whether redistribution
is allowed under the license terms. }
\item{source_url}{String. The URL of the license as listed by the
upstream source. }
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_licenses_list}.
}
\description{
Method for converting openaq_licenses_list to data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
licenses <- list_licenses(as_data_frame = FALSE)
as.data.frame(licenses)
\dontshow{\}) # examplesIf}
}
