% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers.R
\name{as.data.frame.openaq_manufacturers_list}
\alias{as.data.frame.openaq_manufacturers_list}
\title{Method for converting openaq_manufacturers_list to data frame.}
\usage{
\method{as.data.frame}{openaq_manufacturers_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of manufacturers as returned from list_manufacturers.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the manufacturers results, with the following
columns:
\describe{
\item{id}{Numeric. The manufacturers identifier. }
\item{name}{Character. The name of the manufacturer. }
\item{instrument_ids}{Character. A comma delimited list of instrument
identifiers manufactured by the manufacturer.}
\item{instrument_names}{Character. A comma delimited list of instrument
names manufactured by this manufacturer.}
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_manufacturers_list}.
}
\description{
Method for converting openaq_manufacturers_list to data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
manufacturers <- list_manufacturers(as_data_frame = FALSE)
as.data.frame(manufacturers)
\dontshow{\}) # examplesIf}
}
