% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{as.data.frame.openaq_parameters_list}
\alias{as.data.frame.openaq_parameters_list}
\title{Method for converting openaq_parameters_list to data frame.}
\usage{
\method{as.data.frame}{openaq_parameters_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of parameters as returned from list_parameters.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the parameters results, with the following
columns:
\describe{
\item{id}{Numeric. The parameter identifier. }
\item{name}{Character. Name of the parameter.}
\item{units}{Character. The units used by the parameter. }
\item{display_name}{Character. The combined name of the parameter and
units. May be NA if not provided.}
\item{description}{Character. Description of the parameter. May be NA if
not provided.}
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_parameters_list}.
}
\description{
Method for converting openaq_parameters_list to data frame.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parameters <- list_parameters(as_data_frame = FALSE)
as.data.frame(parameters)
\dontshow{\}) # examplesIf}
}
