% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{as.data.frame.openaq_providers_list}
\alias{as.data.frame.openaq_providers_list}
\title{Method for converting openaq_providers_list to data frame.}
\usage{
\method{as.data.frame}{openaq_providers_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of providers as returned from list_providers.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A data frame class of the providers results, with the following
columns:
\describe{
\item{id}{Numeric. The providers identifier.}
\item{name}{Character. The name of the provider.}
\item{source_name}{Factor. The name of the source.}
\item{export_prefix}{Character. Prefixed when exported to file store.}
\item{datetime_added}{POSIXct. Datetime when the provider was first
added to OpenAQ. }
\item{datetime_first}{POSIXct. Datetime of the first measurement value
from the provider.}
\item{datetime_last}{POSIXct. Datetime of the last measurement value
from the provider.}
\item{entities_id}{Numeric. Entities identifier for the provider.}
\item{parameter_ids}{Character. A comma delimited list of parameters
identifier measured by the provider.}
}
The data frame also includes a \code{meta} attribute from the original
\code{openaq_providers_list}.
}
\description{
Method for converting openaq_providers_list to data frame.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
providers <- list_providers(as_data_frame = FALSE)
as.data.frame(providers)
\dontshow{\}) # examplesIf}
}
