% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{disable_rate_limit}
\alias{disable_rate_limit}
\title{Toggles on the RATE_LIMIT environment variable to FALSE}
\usage{
disable_rate_limit()
}
\value{
No return value, called for side effects.
}
\description{
Toggles on the RATE_LIMIT environment variable to FALSE
}
\examples{
disable_rate_limit()

}
