% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{enable_rate_limit}
\alias{enable_rate_limit}
\title{Toggles on the RATE_LIMIT environment variable to TRUE.}
\usage{
enable_rate_limit()
}
\description{
Toggles on the RATE_LIMIT environment variable to TRUE.
}
