% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_countries}
\alias{get_countries}
\title{Get OpenAQ country metadata}
\usage{
get_countries()
}
\value{
A tibble of OpenAQ country metadata.
}
\description{
Returns OpenAQ country metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
