% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{get_instrument}
\alias{get_instrument}
\title{Get a single instrument from the instruments resource.}
\usage{
get_instrument(
  instruments_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  api_key = NULL
)
}
\arguments{
\item{instruments_id}{An integer representing the OpenAQ instruments_id.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a single instrument from the instruments resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
instrument <- get_instrument(42)
\dontshow{\}) # examplesIf}
}
