% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_instruments}
\alias{get_instruments}
\title{Get OpenAQ instrument metadata}
\usage{
get_instruments()
}
\value{
A tibble of OpenAQ instrument metadata.
}
\description{
Returns OpenAQ instrument metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
