% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{get_license}
\alias{get_license}
\title{Get a single license from the licenses resource.}
\usage{
get_license(licenses_id, as_data_frame = TRUE, dry_run = FALSE, api_key = NULL)
}
\arguments{
\item{licenses_id}{An integer representing the OpenAQ licenses_id.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a single license from the licenses resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
license <- get_license(42)
\dontshow{\}) # examplesIf}
}
