% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_licenses}
\alias{get_licenses}
\title{Get OpenAQ license metadata}
\usage{
get_licenses()
}
\value{
A tibble of OpenAQ license metadata.
}
\description{
Returns OpenAQ license metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
