% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_location}
\alias{get_location}
\title{Get a single location from the locations resource.}
\usage{
get_location(
  locations_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  api_key = NULL
)
}
\arguments{
\item{locations_id}{An integer representing the locations_id to request.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or list of results.
}
\description{
Get a single location from the locations resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
location <- get_location(42)
\dontshow{\}) # examplesIf}
}
