% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers.R
\name{get_manufacturer}
\alias{get_manufacturer}
\title{Get a single manufacturer from the manufacturers resource.}
\usage{
get_manufacturer(
  manufacturers_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  api_key = NULL
)
}
\arguments{
\item{manufacturers_id}{An integer representing the OpenAQ manufacturers_id.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a single manufacturer from the manufacturers resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
manufacturer <- get_manufacturer(42)
\dontshow{\}) # examplesIf}
}
