% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_manufacturers}
\alias{get_manufacturers}
\title{Get OpenAQ manufacturer metadata}
\usage{
get_manufacturers()
}
\value{
A tibble of OpenAQ manufacturer metadata.
}
\description{
Returns OpenAQ manufacturer metadata as a tibble. Data are downloaded only
once per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
