% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_owners}
\alias{get_owners}
\title{Get OpenAQ owner metadata}
\usage{
get_owners()
}
\value{
A tibble of OpenAQ owner metadata.
}
\description{
Returns OpenAQ owner metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
