% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_parameters}
\alias{get_parameters}
\title{Get OpenAQ parameter metadata}
\usage{
get_parameters()
}
\value{
A tibble of OpenAQ parameter metadata.
}
\description{
Returns OpenAQ parameter metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
