% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{get_providers}
\alias{get_providers}
\title{Get OpenAQ provider metadata}
\usage{
get_providers()
}
\value{
A tibble of OpenAQ provider metadata.
}
\description{
Returns OpenAQ provider metadata as a tibble. Data are downloaded only once
per R session and then cached for re-use.
}
\details{
For more information about the OpenAQ API, see:
\url{https://docs.openaq.org/}
}
