% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{get_sensor}
\alias{get_sensor}
\title{Get a single sensor from sensors resource.}
\usage{
get_sensor(sensors_id, as_data_frame = TRUE, dry_run = FALSE, api_key = NULL)
}
\arguments{
\item{sensors_id}{An integer representing the OpenAQ sensors_id.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a single sensor from sensors resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sensor <- get_sensor(42)
\dontshow{\}) # examplesIf}
}
