% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{instrument_to_id}
\alias{instrument_to_id}
\title{Convert instrument names to OpenAQ instrument IDs}
\usage{
instrument_to_id(x)
}
\arguments{
\item{x}{Character vector of instrument names.}
}
\value{
Integer vector of OpenAQ instrument IDs.
}
\description{
Matching is case-insensitive. If no match is found, \code{NA} is returned.
}
\examples{
\dontrun{
instrument_to_id(c("BAM 1020", "PurpleAir PA-II", "Unknown"))
}

}
