% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{license_to_id}
\alias{license_to_id}
\title{Convert license names to OpenAQ license IDs}
\usage{
license_to_id(x)
}
\arguments{
\item{x}{Character vector of license names.}
}
\value{
Integer vector of OpenAQ license IDs.
}
\description{
Matching is case-insensitive. If no match is found, \code{NA} is returned.
}
\examples{
\dontrun{
license_to_id(c("CC BY 4.0", "Public Domain", "Unknown"))
}

}
