% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{list_location_measurements}
\alias{list_location_measurements}
\title{Get a list of measurements by locations_id.}
\usage{
list_location_measurements(
  locations_id,
  parameters_ids = NULL,
  data = "measurements",
  rollup = NULL,
  datetime_from = NULL,
  datetime_to = NULL,
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{locations_id}{An integer representing an OpenAQ locations_id}

\item{parameters_ids}{A numeric vector of length 1 or more, containing the
ID(s) of the parameter(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned.}

\item{data}{A string representing the data interval to return, default is
"measurements".}

\item{rollup}{A string representing the aggregation rollup (e.g., 'daily').}

\item{datetime_from}{A POSIXct datetime to filter measurements.}

\item{datetime_to}{A POSIXct datetime to filter measurements.}

\item{order_by}{A string representing the field to order by.}

\item{sort_order}{A string, either "asc" or "desc".}

\item{limit}{An integer representing the number of results per page.}

\item{page}{An integer representing the page number.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a list of measurements by locations_id.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
measurements <- list_location_measurements(locations_id = 3920, "hours")
\dontshow{\}) # examplesIf}
}
