% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{list_parameter_latest}
\alias{list_parameter_latest}
\title{Get the latest measurements by parameters_id.}
\usage{
list_parameter_latest(
  parameters_id,
  datetime_min = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  api_key = NULL
)
}
\arguments{
\item{parameters_id}{An integer representing the OpenAQ parameters_id}

\item{datetime_min}{A POSIXct datetime  specifying the minimum datetime for
filtering results, default is \code{NULL}.}

\item{limit}{An integer specifying the maximum number of results to return,
default is \code{100}.}

\item{page}{An integer specifying the page number for paginated results,
default is \code{1}.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get the latest measurements by parameters_id.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
measurements <- list_parameter_latest(2)
\dontshow{\}) # examplesIf}
}
