% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{manufacturer_to_id}
\alias{manufacturer_to_id}
\title{Convert manufacturer names to OpenAQ manufacturer IDs}
\usage{
manufacturer_to_id(x)
}
\arguments{
\item{x}{Character vector of manufacturer names.}
}
\value{
Integer vector of OpenAQ manufacturer IDs.
}
\description{
Matching is case-insensitive. If no match is found, \code{NA} is returned.
}
\examples{
\dontrun{
manufacturer_to_id(
  c("Met One Instruments", "Thermo Fisher Scientific", "Unknown")
)
}

}
