% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq-package.R
\docType{package}
\name{openaq-package}
\alias{openaq}
\alias{openaq-package}
\title{openaq: Access Air Quality Data from the OpenAQ API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to the OpenAQ API \url{https://openaq.org/}, a platform for real-time and historical air quality data from around the world. Users can retrieve measurement data, metadata for sensors and locations for air quality research and monitoring.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openaq/openaq-r}
  \item \url{https://openaq.github.io/openaq-r/}
  \item Report bugs at \url{https://github.com/openaq/openaq-r/issues}
}

}
\author{
\strong{Maintainer}: Russ Biggs \email{russ@openaq.org} (\href{https://orcid.org/0009-0008-0777-8479}{ORCID})

Authors:
\itemize{
  \item Christian Parker \email{chris@talloaks.io}
}

Other contributors:
\itemize{
  \item Francesca Mills [contributor]
  \item OpenAQ \email{dev@openaq.org} [copyright holder, funder]
}

}
\keyword{internal}
