% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_tables.R
\name{owner_to_id}
\alias{owner_to_id}
\title{Convert owner names to OpenAQ owner IDs}
\usage{
owner_to_id(x)
}
\arguments{
\item{x}{Character vector of owner names.}
}
\value{
Integer vector of OpenAQ owner IDs.
}
\description{
Matching is case-insensitive. If no match is found, \code{NA} is returned.
}
\examples{
\dontrun{
owner_to_id(
  c("Environmental Defense Fund", "Carnegie Mellon University")
)
}

}
