% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{plot.openaq_locations_data.frame}
\alias{plot.openaq_locations_data.frame}
\title{Helper for plotting locations on map.}
\usage{
\method{plot}{openaq_locations_data.frame}(loc, database = "world", ...)
}
\arguments{
\item{loc}{A data frame of locations results.}

\item{database}{the maps package database of geographic boundaries to use,
defaults to "world".}

\item{...}{Other options passed on to base::plot().}
}
\description{
Helper for plotting locations on map.
}
\examples{
\dontrun{
df <- list_locations(limit = 100)
plot(df, pch = 19, col = df$provider)
}
}
