% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{plot.openaq_measurements_data.frame}
\alias{plot.openaq_measurements_data.frame}
\title{Helper for plotting measurements}
\usage{
\method{plot}{openaq_measurements_data.frame}(df, ...)
}
\arguments{
\item{df}{A measurements data frame.}
}
\description{
Helper for plotting measurements
}
\examples{
\dontrun{
meas <- list_sensor_measurements(23707, limit = 500, as_data_frame = FALSE)
plot(meas)
}
}
