% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the API key value.}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{A string value for the API key to set.}
}
\description{
A helper function to set the OPENAQ_API_KEY environment variable.
}
\examples{
set_api_key("my-super-secret-openaq-api-key-1234")

}
