% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_base_url}
\alias{set_base_url}
\title{Sets base URL environment variable}
\usage{
set_base_url(base_url)
}
\arguments{
\item{base_url}{A string.}
}
\description{
A helper function to set the OPENAQR_BASE_URL environment variable. This is
to override the default URL for testing and custom instance of the API. This
function is generally not used by most users except in extraordinary cases.
}
\examples{
set_base_url("https://example.com")

}
