% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{as.data.frame.openaq_countries_list}
\alias{as.data.frame.openaq_countries_list}
\title{Method for converting openaq_countries_list to data frame.}
\usage{
\method{as.data.frame}{openaq_countries_list}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A list of countries as returned from list_countries}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of R's
base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Method for converting openaq_countries_list to data frame.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
countries <- list_countries()
openaq_countries_list.as.data.frame(countries)
\dontshow{\}) # examplesIf}
}
