% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{as.data.frame.openaq_latest_list}
\alias{as.data.frame.openaq_latest_list}
\title{Method for converting openaq_latest_list to data frame.}
\usage{
\method{as.data.frame}{openaq_latest_list}(data, ...)
}
\arguments{
\item{data}{A list of countries as returned from list_instruments}
}
\description{
Method for converting openaq_latest_list to data frame.
}
\examples{
\dontrun{
instruments <- list_instruments(as_data_frame = FALSE)
openaq_latest_list.as.data.frame(instruments)
}
}
