% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{as.data.frame.openaq_locations_list}
\alias{as.data.frame.openaq_locations_list}
\title{Method for converting to data frame}
\usage{
\method{as.data.frame}{openaq_locations_list}(data, ...)
}
\arguments{
\item{data}{A list of locations as returned from list_locations.}

\item{...}{Other options.}
}
\description{
Method for converting to data frame
}
\examples{
\dontrun{
loc <- list_locations()
write.csv(loc)
}
}
