% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{as.data.frame.openaq_measurements_list}
\alias{as.data.frame.openaq_measurements_list}
\title{Method for converting openaq_measurements_list to data frame.}
\usage{
\method{as.data.frame}{openaq_measurements_list}(data, ...)
}
\arguments{
\item{data}{A list of measurements as returned from list_sensor_measurements.}

\item{...}{Other options.}
}
\description{
Method for converting openaq_measurements_list to data frame.
}
\examples{
\dontrun{
meas <- list_sensor_measurements(23707, limit = 500, as_data_frame = FALSE)
openaq_measurements_list.as.data.frame(meas)
}
}
