% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{as.data.frame.openaq_providers_list}
\alias{as.data.frame.openaq_providers_list}
\title{Method for converting openaq_providers_list to data frame.}
\usage{
\method{as.data.frame}{openaq_providers_list}(x, row.names=NULL, optional=FALSE, ...)
}
\arguments{
\item{x}{A list of countries as returned from list_providers.}

\item{row.names}{‘NULL’ or a character vector giving the row names for
  the data frame.  Missing values are not allowed.}

\item{optional}{logical. Should we check names?}

\item{...}{Other options.}
}
\description{
Method for converting openaq_providers_list to data frame.
}
\examples{
\dontrun{
providers <- list_providers(as_data_frame = FALSE)
openaq_instruments_list.as.data.frame(providers)
}
}
