% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{as.data.frame.openaq_sensors_list}
\alias{as.data.frame.openaq_sensors_list}
\title{Method for converting openaq_sensors_list to data frame.}
\usage{
\method{as.data.frame}{openaq_sensors_list}(data, ...)
}
\arguments{
\item{data}{A list of sensors as returned from get_sensor}

\item{...}{Other options.}
}
\description{
Method for converting openaq_sensors_list to data frame.
}
\examples{
\dontrun{
sensor <- get_sensor(as_frame_frame = FALSE)
openaq_sensors_list.as.data.frame(sensor)
}
}
