% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{as.data.frame.openaq_sensors_list}
\alias{as.data.frame.openaq_sensors_list}
\title{Method for converting openaq_sensors_list to data frame.}
\usage{
\method{as.data.frame}{openaq_sensors_list}(x, row.names=NULL, optional=FALSE, ...)
}
\arguments{
\item{x}{A list of sensors as returned from get_sensor}

\item{row.names}{‘NULL’ or a character vector giving the row names for
  the data frame.  Missing values are not allowed.}

\item{optional}{logical. Should we check names?}

\item{...}{Other options.}
}
\description{
Method for converting openaq_sensors_list to data frame.
}
\examples{
\dontrun{
sensor <- get_sensor(as_frame_frame = FALSE)
openaq_sensors_list.as.data.frame(sensor)
}
}
