% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_country}
\alias{get_country}
\title{Get a single country from countries resource.}
\usage{
get_country(
  countries_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{countries_id}{An integer representing the OpenAQ countries_id.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame or list of the results.
}
\description{
Get a single country from countries resource.
}
\examples{
\dontrun{
country <- get_country(42)
}

}
