% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{get_license}
\alias{get_license}
\title{get a single license from the licenses resource.}
\usage{
get_license(
  licenses_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{licenses_id}{An integer.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame.
}
\description{
get a single license from the licenses resource.
}
\examples{
\dontrun{
license <- get_license(42)
}

}
