% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_location}
\alias{get_location}
\title{Get a single location from the locations resource.}
\usage{
get_location(
  locations_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{locations_id}{An integer representing the locations_id to request.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame or list of results.
}
\description{
Get a single location from the locations resource.
}
\examples{
\dontrun{
location <- get_location(42)
}

}
