% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licenses.R
\name{list_licenses}
\alias{list_licenses}
\title{Get a list of licenses from the licenses resource.}
\usage{
list_licenses(
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  api_key = NULL
)
}
\arguments{
\item{order_by}{A character string specifying the field to order results by.}

\item{sort_order}{A character string specifying sort direction, either \code{"asc"} or \code{"desc"}.}

\item{limit}{An integer specifying the maximum number of results to return, default is \code{100}.}

\item{page}{An integer specifying the page number for paginated results, default is \code{1}.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a list of licenses from the licenses resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
licenses <- list_licenses()
\dontshow{\}) # examplesIf}
}
