% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{list_location_latest}
\alias{list_location_latest}
\title{Get the latest measurements by locations_id.}
\usage{
list_location_latest(
  locations_id,
  datetime_min = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{locations_id}{An integer.}

\item{datetime_min}{A string.}

\item{limit}{An integer.}

\item{page}{An integer.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
\code{FALSE}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get the latest measurements by locations_id.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
measurements <- list_location_latest(2178)
\dontshow{\}) # examplesIf}
}
