% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{list_location_measurements}
\alias{list_location_measurements}
\title{Get a list of measurements by locations_id.}
\usage{
list_location_measurements(
  locations_id,
  parameters_ids = NULL,
  data = "measurements",
  rollup = NULL,
  datetime_from = NULL,
  datetime_to = NULL,
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{locations_id}{An integer representing an OpenAQ locations_id}

\item{parameters_ids}{A numeric vector of length 1 or more, containing the
ID(s) of the parameter(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned.}

\item{data}{A string a data interval to return, defaults to "measurements".}

\item{rollup}{A string.}

\item{datetime_from}{A POSIXct datetime to filter measurements}

\item{datetime_to}{A POSIXct datetime to filter measurements}

\item{order_by}{A string.}

\item{sort_order}{A string.}

\item{limit}{An integer.}

\item{page}{An integer.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a list of measurements by locations_id.
}
\examples{
\dontrun{
measurements <- list_sensor_measurements(3920, "hours")
}

}
