% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{list_locations}
\alias{list_locations}
\title{Get a list of locations from the locations resource.}
\usage{
list_locations(
  bbox = NULL,
  coordinates = NULL,
  radius = NULL,
  providers_id = NULL,
  parameters_id = NULL,
  owner_contacts_id = NULL,
  manufacturers_id = NULL,
  licenses_id = NULL,
  monitor = NULL,
  mobile = NULL,
  instruments_id = NULL,
  iso = NULL,
  countries_id = NULL,
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{bbox}{Named numeric vector with four coordinates in form X minimum,
Y mininum, X maximum, Y maximum, named values must be \code{xmin}, \code{ymin}, \code{ymax}
, \code{xmax}. Defaults to \code{NULL}.}

\item{coordinates}{Named numeric vector with two numeric WGS84 (EPSG:4326)
geographic coordinates, with named values \code{latitude} and \code{longitude}.
Represents the central point to be used in conjunction with the radius
parameter for geographic search. Defaults to \code{NULL}.}

\item{radius}{An integer for the number of meters to search around the
\code{coordinates} parameter for filtering locations within the radius. Defaults
to \code{NULL}.}

\item{providers_id}{A numeric vector of length 1 or more, containing the
ID(s) of the provider(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{parameters_id}{A numeric vector of length 1 or more, containing the
ID(s) of the parameter(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{owner_contacts_id}{A numeric vector of length 1 or more, containing the
ID(s) of the owners(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{manufacturers_id}{A numeric vector of length 1 or more, containing the
ID(s) of the manufacturers(s) to use for filtering results. If multiple IDs
are provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{licenses_id}{A numeric vector of length 1 or more, containing the
ID(s) of the license(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{monitor}{A logical to filter results to regulatory monitors (TRUE) or
air sensors (FALSE), both are included if NULL. Defaults to \code{NULL}.}

\item{mobile}{A logical to filter results to mobile (TRUE) or stationary
(FALSE) location, both are included if NULL. Defaults to \code{NULL}.}

\item{instruments_id}{A numeric vector of length 1 or more, containing the
ID(s) of the instrument(s) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{iso}{An ISO 3166-1 alpha-2 string of the country to filter the results.}

\item{countries_id}{A numeric vector of length 1 or more, containing the
ID(s) of the country(ies) to use for filtering results. If multiple IDs are
provided, results matching any of the IDs will be returned. Defaults to
\code{NULL}.}

\item{order_by}{A string. Defaults to \code{NULL}.}

\item{sort_order}{A string. Defaults to \code{NULL}.}

\item{limit}{An integer to limit the number of results per page. Defaults to
\code{NULL}.}

\item{page}{An integer for the page number for paginating through results.
Defaults to \code{NULL}.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list. Defaults to \code{TRUE}}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
\code{FALSE.}}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers. Defaults to \code{FALSE.}}

\item{api_key}{A valid OpenAQ API key string. Defaults to \code{NULL}.}
}
\value{
A data frame or list of results.
}
\description{
Get a list of locations from the locations resource.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
locations <- list_locations()
\dontshow{\}) # examplesIf}
}
