% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{list_manufacturer_instruments}
\alias{list_manufacturer_instruments}
\title{get a list of manufacturer instruments from the instruments resource.}
\usage{
list_manufacturer_instruments(
  manufacturers_id,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{manufacturers_id}{An integer.}
}
\value{
A data frame.
}
\description{
get a list of manufacturer instruments from the instruments resource.
}
\examples{
\dontrun{
instruments <- list_manufacturer_instruments(42)
}

}
