% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers.R
\name{list_manufacturers}
\alias{list_manufacturers}
\title{get a list of manufacturers from the manufacturers resource.}
\usage{
list_manufacturers(
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{order_by}{A string.}

\item{sort_order}{A string.}

\item{limit}{An integer.}

\item{page}{An integer.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame or list of results.
}
\description{
get a list of manufacturers from the manufacturers resource.
}
\examples{
\dontrun{
manufacturers <- list_manufacturers()
}

}
