% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{list_providers}
\alias{list_providers}
\title{get a list of providers from the providers resource.}
\usage{
list_providers(
  order_by = NULL,
  sort_order = NULL,
  limit = 100,
  page = 1,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{order_by}{A string.}

\item{sort_order}{A string.}

\item{limit}{An integer.}

\item{page}{An integer.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list defaults to TRUE.}

\item{dry_run}{A logical for toggling a dry run of the request, defaults to
FALSE.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, defaults to FALSE.}

\item{api_key}{A valid OpenAQ API key string, defaults to NULL.}
}
\value{
A data frame.
}
\description{
get a list of providers from the providers resource.
}
\examples{
\dontrun{
providers <- list_providers()
}

}
