% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{list_sensor_measurements}
\alias{list_sensor_measurements}
\title{Get a list of measurements by sensors_id.}
\usage{
list_sensor_measurements(
  sensors_id,
  data = "measurements",
  rollup = NULL,
  datetime_from = NULL,
  datetime_to = NULL,
  order_by = NULL,
  sort_order = NULL,
  limit = NULL,
  page = NULL,
  as_data_frame = TRUE,
  dry_run = FALSE,
  rate_limit = FALSE,
  api_key = NULL
)
}
\arguments{
\item{sensors_id}{An integer representing an OpenAQ sensors_id.}

\item{data}{A string a data interval to return, default is "measurements".}

\item{rollup}{A string representing the aggregation rollup, default is \code{NULL}.}

\item{datetime_from}{A POSIXct datetime (when \code{data} is \code{"measurements"} or
\code{"hours"}) or a Date (when \code{data} is \code{"days"} or larger) to filter from,
default is \code{NULL}.}

\item{datetime_to}{A POSIXct datetime (when \code{data} is \code{"measurements"} or
\code{"hours"}) or a Date (when \code{data} is \code{"days"} or larger) to filter to,
default is \code{NULL}.}

\item{order_by}{A string representing the field to order by, default is \code{NULL}.}

\item{sort_order}{A string, either "asc" or "desc", default is \code{NULL}.}

\item{limit}{An integer representing the number of results per page.}

\item{page}{An integer representing the page number.}

\item{as_data_frame}{A logical for toggling whether to return results as
data frame or list, default is \code{TRUE}.}

\item{dry_run}{A logical for toggling a dry run of the request, default is
\code{FALSE}.}

\item{rate_limit}{A logical for toggling automatic rate limiting based on
rate limit headers, default is \code{FALSE}.}

\item{api_key}{A valid OpenAQ API key string, default is \code{NULL}.}
}
\value{
A data frame or a list of the results.
}
\description{
Get a list of measurements by sensors_id.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
measurements <- list_sensor_measurements(3920, "hours")
\dontshow{\}) # examplesIf}
}
