% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq-package.R
\docType{package}
\name{openaq-package}
\alias{openaq}
\alias{openaq-package}
\title{openaq: OpenAQ API Client}
\description{
Allows access to the version 3 of the OpenAQ API \url{https://openaq.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openaq/openaq-r}
  \item Report bugs at \url{https://github.com/openaq/openaq-r/issues}
}

}
\author{
\strong{Maintainer}: Russ Biggs \email{russ@openaq.org}

Authors:
\itemize{
  \item Christian Parker \email{chris@talloaks.io}
}

Other contributors:
\itemize{
  \item OpenAQ \email{dev@openaq.org} [copyright holder, funder]
}

}
\keyword{internal}
