% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{plot.openaq_locations_data.frame}
\alias{plot.openaq_locations_data.frame}
\title{helper for plotting locations on map}
\usage{
\method{plot}{openaq_locations_data.frame}(loc, database = "world", ...)
}
\description{
helper for plotting locations on map
}
\examples{
\dontrun{
df <- list_locations(limit = 100)
plot(df, pch = 19, col = df$provider)
}
}
