% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{plot.openaq_locations_data.frame}
\alias{plot.openaq_locations_data.frame}
\title{Helper for plotting locations on map.}
\usage{
\method{plot}{openaq_locations_data.frame}(x, y = NULL, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a single
plotting structure, function or any R object with a plot method can be provided.}

\item{y}{the y coordinates of points in the plot, optional if x is an
appropriate structure.}

\item{...}{Other options passed on to base::plot().}
}
\description{
Plots location coordinates. If the \code{maps} package is installed, a world
boundary overlay is added. Install with \code{install.packages("maps")}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- list_locations(limit = 100)
plot(df, pch = 19, col = df$provider_name)
\dontshow{\}) # examplesIf}
}
