% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{plot.openaq_locations_list}
\alias{plot.openaq_locations_list}
\title{Helper for plotting locations from list.}
\usage{
\method{plot}{openaq_locations_list}(loc, database = "world", ...)
}
\arguments{
\item{loc}{A list of locations results.}

\item{database}{the maps package database of geographic boundaries to use,
defaults to "world".}

\item{...}{Other options passed on to base::plot().}
}
\description{
Helper for plotting locations from list.
}
\examples{
\dontrun{
loc <- list_locations(limit = 6, as_data_frame = FALSE)
plot(loc, pch = 19, col = 2)
}
}
