% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{plot.openaq_measurements_list}
\alias{plot.openaq_measurements_list}
\title{Helper for plotting measurements from list}
\usage{
\method{plot}{openaq_measurements_list}(x, y = NULL, ...)
}
\arguments{
\item{x}{A list of measurements results.}

\item{y}{Other data}

\item{...}{Other options to be passed on to base::plot().}
}
\description{
Helper for plotting measurements from list
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
meas <- list_sensor_measurements(23707, limit = 500)
plot(meas)
\dontshow{\}) # examplesIf}
}
