% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{plot.openaq_measurements_list}
\alias{plot.openaq_measurements_list}
\title{Helper for plotting measurements from list}
\usage{
\method{plot}{openaq_measurements_list}(meas, ...)
}
\description{
Helper for plotting measurements from list
}
\examples{
\dontrun{
meas <- list_sensor_measurements(23707, limit = 500, as_data_frame = FALSE)
plot(meas)
}
}
